Improve tests introducing stopWhen method (#3)
Before this refactor tests were stopped after an arbitrary delay.
This delay was too low under certain conditions (for example under
docker on macOS).

So instead of creating a configurable delay mechanism this commit
introduces a stopWhen logic which stops the event loop when a given
condition is reached. If the condition is not reached within a
certain timeout the test fails.